重构：性能优化 + 动态高度 + 主账号 + 代码维护
- 共享 reqwest::Client 连接池，API Key 内存缓存减少 Keychain IPC
- refresh_all 返回完整 QuotaData，消除前端冗余请求
- ResizeObserver 驱动动态窗口高度，内容自适应，上限屏幕高度
- overflow-y: overlay 消除滚动条抖动
- 主账号功能：星标切换、托盘百分比跟随主账号、删除自动提升
- 多账号同时展开，TOKENS_LIMIT badge 显示
- 提取公共 Toggle 组件，修复暗色模式颜色硬编码
- 移除未使用结构体/空文件/ZhipuClient::new()

+++ src-tauri/src/alert/notifier.rs
@@ -30,19 +30,19 @@ pub fn check_and_notify(
                 .limits
                 .iter()
                 .find(|l| l.limit_type == "TIME_LIMIT")
-                .map(|l| l.percentage as f64)
+                .map(|l| l.percentage)
                 .unwrap_or(0.0),
             "weekly" => quota
                 .limits
                 .iter()
                 .find(|l| l.limit_type == "TOKENS_LIMIT")
-                .map(|l| l.percentage as f64)
+                .map(|l| l.percentage)
                 .unwrap_or(0.0),
             "mcp_monthly" => quota
                 .limits
                 .iter()
                 .find(|l| l.limit_type == "MCP_MONTHLY")
-                .map(|l| l.percentage as f64)
+                .map(|l| l.percentage)
                 .unwrap_or(0.0),
             _ => continue,
         };
